requirements.txt
blinker==1.6.2
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.5
cycler==0.11.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
fonttools==4.31.2
greenlet==2.0.2
gunicorn==21.2.0
idna==3.3
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.0
MarkupSafe==2.1.3
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.3
packaging==23.1
Pillow==9.0.1
PyJWT==2.8.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.28.1
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==1.26.11
watchdog==3.0.0
Werkzeug==2.3.6
zipp==3.16.2
